Code/requirements.txt
requests
flask
flask-restful
celery
Redis
gunicorn==20.1.0
google-cloud-tasks==2.7.1
google-api-python-client
google-cloud-speech
google-cloud-storage